+++ src/eval/requirement/detection/detection.cpp
@@ -41,7 +41,7 @@ namespace EvaluationRequirement
 Detection::Detection(const std::string& name, 
                      const std::string& short_name, 
                      const std::string& group_name,
-                     float prob, 
+                     double prob, 
                      COMPARISON_TYPE prob_check_type, 
                      EvaluationManager& eval_man,
                      float update_interval_s,
