+++ src/eval/requirement/identification/false.cpp
@@ -34,7 +34,7 @@ namespace EvaluationRequirement
 
 IdentificationFalse::IdentificationFalse(
         const std::string& name, const std::string& short_name, const std::string& group_name,
-        float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man,
+        double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man,
         bool require_all_false, bool use_mode_a, bool use_ms_ta, bool use_ms_ti)
     : ProbabilityBase(name, short_name, group_name, prob, prob_check_type, false, eval_man),
       require_all_false_(require_all_false),
