+++ src/eval/requirement/mode_c/correct.h
@@ -29,7 +29,7 @@ class ModeCCorrect : public ProbabilityBase
 public:
     ModeCCorrect(
             const std::string& name, const std::string& short_name, const std::string& group_name,
-            float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man, float max_distance_ft);
+            double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man, float max_distance_ft);
 
     virtual std::shared_ptr<EvaluationRequirementResult::Single> evaluate (
             const EvaluationTargetData& target_data, std::shared_ptr<Base> instance,
