+++ src/eval/requirement/mode_c/correct_period.h
@@ -35,7 +35,7 @@ public:
             const std::string& name, 
             const std::string& short_name, 
             const std::string& group_name,
-            float prob, 
+            double prob, 
             COMPARISON_TYPE prob_check_type,
             EvaluationManager& eval_man,
             float update_interval_s,
